Fix when brew and docker are not installed.

+++ .bash_profile
@@ -4,7 +4,7 @@
 
 # https://superuser.com/q/267771
 # Put first to ensure the signal handler is set up before sending the signal.
-trap 'source $(brew --prefix)/share/bash-completion/bash_completion; trap USR1' USR1
+[ -x brew ] && trap 'source $(brew --prefix)/share/bash-completion/bash_completion; trap USR1' USR1
 
 # Brew takes precedence.
 export PATH=/usr/local/bin:/usr/local/sbin:$PATH
@@ -21,4 +21,4 @@ if [[ -r ~/.bashrc ]]; then
     . ~/.bashrc
 fi
 
-builtin kill -USR1 $$ & disown
+[ -x brew ] && builtin kill -USR1 $$ & disown
